Change from float to double for P5. Allows use of System.Numerics types.

+++ Helpers/P5/Models/Bezier.cs
@@ -1,3 +1,5 @@
+using System.Numerics;
+
 namespace ExoKomodo.Helpers.P5.Models
 {
     public struct Bezier
@@ -6,18 +8,18 @@ namespace ExoKomodo.Helpers.P5.Models
 
         #region Constructors
         public Bezier(
-            double x1,
-            double y1,
-            double z1,
-            double x2,
-            double y2,
-            double z2,
-            double x3,
-            double y3,
-            double z3,
-            double x4,
-            double y4,
-            double z4
+            float x1,
+            float y1,
+            float z1,
+            float x2,
+            float y2,
+            float z2,
+            float x3,
+            float y3,
+            float z3,
+            float x4,
+            float y4,
+            float z4
         ) : this(
             new Vector3(x1, y1, z1),
             new Vector3(x2, y2, z2),
